Split docs/tests.py into different test files.
